If a post has a description, use it in the XML data
I noticed that the system was putting the entire post body in the description.  It is more desirable to place just the description of the post in there.  You can now add the description: yaml markup to your posts to have a more clean feed XML.  Alternatively, if you do not specify a post description, the page title will be used (instead of the post body).