Add syndication module to provide easy configuration for update hints.
This commit adds the `<sy:updatePeriod>` and `<sy:updateFrequency>`
elements to all feeds. These elements are defined in the RSS 1.0
Syndication Modules specification, described here:

    http://web.resource.org/rss/1.0/modules/syndication/

The purpose of these elements is to provide hints for automated feed
consumers regarding how often to expect updates, which assists in
scheduling polling times so as not to overly tax a given server.

The templates use the default values intentionally so as to avoid
changing the default behavior of the feeds as-is. Configuration options
have been provided to make changing the values easier, since all that's
needed is an edit to `_config.yml`.